Attempt to GJK collision
Still a bit buggy, but mostly works the way it should. Try it out by
running testGJK. Right now it doesn't detect collision "face-on", but
try moving around a bit to see how it works. A later step is to adjust
the algorithm to also calculate the collision point, rather than just a
boolean "hit or no hit".